+++ frontend/src/main.ts
@@ -7,10 +7,13 @@ import 'primeicons/primeicons.css'
 import App from './App.vue'
 import router from './router'
 
+import ToastService from 'primevue/toastservice'
+
 const app = createApp(App)
 
 app.use(createPinia())
 app.use(router)
+app.use(ToastService)
 app.use(PrimeVue, {
   theme: {
     preset: Aura,